reorg optimizer (#6016)
